모의고사 점수확인 할때 백분위 더블형으로 바꿔줌

+++ School_Project/src/kosta/model/MockResearchScoreDetailList.java
@@ -8,18 +8,18 @@ public class MockResearchScoreDetailList implements Serializable {
 	private String researchSubjectId1;
 	private int researchSubjectOriginalScore1;
 	private int researchSubjectStandardScore1;
-	private int researchSubjectPercentile1;
+	private double researchSubjectPercentile1;
 	private String researchSubjectId2;
 	private int researchSubjectOriginalScore2;
 	private int researchSubjectStandardScore2;
-	private int researchSubjectPercentile2;
+	private double researchSubjectPercentile2;
 	
 	public MockResearchScoreDetailList(){}
 
 	public MockResearchScoreDetailList(String memberId, String mockId, String researchSubjectId1,
-			int researchSubjectOriginalScore1, int researchSubjectStandardScore1, int researchSubjectPercentile1,
+			int researchSubjectOriginalScore1, int researchSubjectStandardScore1, double researchSubjectPercentile1,
 			String researchSubjectId2, int researchSubjectOriginalScore2, int researchSubjectStandardScore2,
-			int researchSubjectPercentile2) {
+			double researchSubjectPercentile2) {
 		super();
 		this.memberId = memberId;
 		this.mockId = mockId;
@@ -73,11 +73,11 @@ public class MockResearchScoreDetailList implements Serializable {
 		this.researchSubjectStandardScore1 = researchSubjectStandardScore1;
 	}
 
-	public int getResearchSubjectPercentile1() {
+	public double getResearchSubjectPercentile1() {
 		return researchSubjectPercentile1;
 	}
 
-	public void setResearchSubjectPercentile1(int researchSubjectPercentile1) {
+	public void setResearchSubjectPercentile1(double researchSubjectPercentile1) {
 		this.researchSubjectPercentile1 = researchSubjectPercentile1;
 	}
 
@@ -105,13 +105,14 @@ public class MockResearchScoreDetailList implements Serializable {
 		this.researchSubjectStandardScore2 = researchSubjectStandardScore2;
 	}
 
-	public int getResearchSubjectPercentile2() {
+	public double getResearchSubjectPercentile2() {
 		return researchSubjectPercentile2;
 	}
 
-	public void setResearchSubjectPercentile2(int researchSubjectPercentile2) {
+	public void setResearchSubjectPercentile2(double researchSubjectPercentile2) {
 		this.researchSubjectPercentile2 = researchSubjectPercentile2;
 	}
+
